Create AIML
The field of ophthalmology, dedicated to the care and health of our eyes, is embracing the transformative power of artificial intelligence (AI). AI encompasses a range of techniques that enable computers to mimic human capabilities like learning and problem-solving. In ophthalmology, this translates to various applications that hold immense promise for:
Enhanced diagnosis
Improved screening
Streamlined workflows
Personalized treatment planning
Designing a Eye disease detection app so that the diagnosis can be easily accessible to all.
This will lead to early detection of disease & hence can be dealt in early stage.